+++ .wpmrc.js
@@ -1,5 +1,5 @@
 module.exports = {
-    allowBranch: ['master', 'release-auto-*'],
+    allowBranch: ['master', 'v15.*', 'v16.*', 'v17.*', 'release-auto-*'],
     bumpFiles: ['package.json', 'package-lock.json', 'packages/gantt/package.json'],
     skip: {
         confirm: true