+++ packages/gantt/src/gantt-dom.service.ts
@@ -40,16 +40,25 @@ export class GanttDomService implements OnDestroy {
 
     public linksOverlay: Element;
 
+    private mainFooter: Element;
+
+    private mainScrollbar: Element;
+
     private unsubscribe$ = new Subject<void>();
 
     constructor(private ngZone: NgZone, @Inject(PLATFORM_ID) private platformId: string) {}
 
     private monitorScrollChange() {
+        const scrollObservers = [
+            fromEvent(this.mainContainer, 'scroll', passiveListenerOptions),
+            fromEvent(this.sideContainer, 'scroll', passiveListenerOptions)
+        ];
+
+        this.mainFooter && scrollObservers.push(fromEvent(this.mainFooter, 'scroll', passiveListenerOptions));
+        this.mainScrollbar && scrollObservers.push(fromEvent(this.mainScrollbar, 'scroll', passiveListenerOptions));
+
         this.ngZone.runOutsideAngular(() =>
-            merge(
-                fromEvent(this.mainContainer, 'scroll', passiveListenerOptions),
-                fromEvent(this.sideContainer, 'scroll', passiveListenerOptions)
-            )
+            merge(...scrollObservers)
                 .pipe(takeUntil(this.unsubscribe$))
                 .subscribe((event) => {
                     this.syncScroll(event);
@@ -59,11 +68,22 @@ export class GanttDomService implements OnDestroy {
 
     private syncScroll(event: Event) {
         const target = event.currentTarget as HTMLElement;
-        this.calendarHeader.scrollLeft = this.mainContainer.scrollLeft;
-        this.calendarOverlay.scrollLeft = this.mainContainer.scrollLeft;
+        const classList = target.classList;
 
-        this.sideContainer.scrollTop = target.scrollTop;
-        this.mainContainer.scrollTop = target.scrollTop;
+        if (!classList.contains('gantt-side-container')) {
+            this.mainContainer.scrollLeft = target.scrollLeft;
+            this.calendarHeader.scrollLeft = target.scrollLeft;
+            this.calendarOverlay.scrollLeft = target.scrollLeft;
+            this.mainScrollbar && (this.mainScrollbar.scrollLeft = target.scrollLeft);
+            this.mainFooter && (this.mainFooter.scrollLeft = target.scrollLeft);
+            if (classList.contains('gantt-main-container')) {
+                this.sideContainer.scrollTop = target.scrollTop;
+                this.mainContainer.scrollTop = target.scrollTop;
+            }
+        } else {
+            this.sideContainer.scrollTop = target.scrollTop;
+            this.mainContainer.scrollTop = target.scrollTop;
+        }
     }
 
     private disableBrowserWheelEvent() {
@@ -92,10 +112,11 @@ export class GanttDomService implements OnDestroy {
         this.container = this.root.getElementsByClassName('gantt-container')[0];
         this.sideContainer = this.root.getElementsByClassName('gantt-side-container')[0];
         this.mainContainer = this.root.getElementsByClassName('gantt-main-container')[0];
+        this.mainScrollbar = this.root.getElementsByClassName('gantt-main-scrollbar')[0];
+        this.mainFooter = this.root.getElementsByClassName('gantt-container-footer')[0];
         this.verticalScrollContainer = this.root.getElementsByClassName('gantt-scroll-container')[0];
         const mainItems = this.mainContainer.getElementsByClassName('gantt-main-items')[0];
         const mainGroups = this.mainContainer.getElementsByClassName('gantt-main-groups')[0];
-
         this.mainItems = mainItems || mainGroups;
         this.calendarHeader = this.root.getElementsByClassName('gantt-calendar-header')[0];
         this.calendarOverlay = this.root.getElementsByClassName('gantt-calendar-grid')[0];
@@ -109,9 +130,13 @@ export class GanttDomService implements OnDestroy {
      * to run the change detection if needed.
      */
     getViewerScroll(options?: AddEventListenerOptions): Observable<ScrollEvent> {
+        const scrollObservers = [fromEvent(this.mainContainer, 'scroll', options)];
+        this.mainFooter && scrollObservers.push(fromEvent(this.mainFooter, 'scroll', options));
+        this.mainScrollbar && scrollObservers.push(fromEvent(this.mainScrollbar, 'scroll', options));
+
         return new Observable<ScrollEvent>((subscriber) =>
             this.ngZone.runOutsideAngular(() =>
-                fromEvent(this.mainContainer, 'scroll', options)
+                merge(...scrollObservers)
                     .pipe(
                         map(() => this.mainContainer.scrollLeft),
                         pairwise(),
@@ -151,6 +176,8 @@ export class GanttDomService implements OnDestroy {
             this.mainContainer.scrollLeft = scrollLeft > scrollThreshold ? scrollLeft : 0;
             this.calendarHeader.scrollLeft = this.mainContainer.scrollLeft;
             this.calendarOverlay.scrollLeft = this.mainContainer.scrollLeft;
+            this.mainScrollbar && (this.mainScrollbar.scrollLeft = this.mainContainer.scrollLeft);
+            this.mainFooter && (this.mainFooter.scrollLeft = this.mainContainer.scrollLeft);
         }
     }